Make it faster
It turns out that keystone reinitializes the drivers several times for
one single action, as opposed to once after restarting. Our
initialization was making LDAP calls for every single user in the user
map, and that was happening multiple times for every single action. This
patch speeds things up by:

1) No longer making LDAP calls in __init__, since we can validate the
existence of a user later.
2) Fixing list_role_assignments so that if a user ID is provided it only
deals with that user and does not bother with the rest of the list.